requirements.txt
openai==0.28
pandas==2.2.3
python-dotenv==1.1.0